Backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
PyMuPDF
faiss-cpu
numpy